+++ far/edit.cpp
@@ -423,9 +423,9 @@ void Edit::FastShow(const ShowInfo* Info)
 
 			Global->ScrBuf->ApplyColor(
 				{
-					std::min(m_Where.left + TabSelStart, static_cast<int>(m_Where.right + 1)),
+					std::min(m_Where.left + TabSelStart, m_Where.right + 1),
 					m_Where.top,
-					std::min(m_Where.left + TabSelEnd - 1, static_cast<int>(m_Where.right + 1)),
+					std::min(m_Where.left + TabSelEnd - 1, m_Where.right + 1),
 					m_Where.top
 				},
 				GetSelectedColor()
